+++ app/src/main/java/com/niloy/data/local/AppDatabase.kt
@@ -22,7 +22,7 @@ import com.niloy.data.local.entity.TaskOccurrenceEntity
         SettingEntity::class,
         AppClassificationEntity::class
     ],
-    version = 2,
+    version = 3,
     exportSchema = false
 )
 @TypeConverters(Converters::class)